# resources/docs/openapi/openapi.yaml
...
openapi: 3.0.3
info:
    title: OrigAdmin API
    description: A lightweight, flexible, elegant and full-featured RBAC scaffolding backend management project.
    contact:
        name: GodCong
        url: https://github.com/origadmin
        email: waitforadding@gmail.com
    license:
        name: MIT
        url: https://github.com/origadmin/backend/blob/master/LICENSE
    version: Version from annotation
servers:
    - url: http://localhost:10080
    - url: https://localhost:10080
paths:
...
    /sys/auth/resources:
        get:
            tags:
                - AuthService
            operationId: AuthService_ListAuthResources
            parameters:
                - name: page_size
                  in: query
                  description: The maximum number of Auths to return.
                  schema:
                    type: integer
                    format: int32
                - name: page_token
                  in: query
                  description: The next_page_token value returned from a previous List request, if any.
                  schema:
                    type: string
                - name: current
                  in: query
                  description: The current page number.
                  schema:
                    type: integer
                    format: int32
                - name: no_paging
                  in: query
                  description: The no_paging is used to disable pagination.
                  schema:
                    type: boolean
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.system.ListAuthResourcesResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
...
    /sys/resources:
        get:
            tags:
                - ResourceService
            operationId: ResourceService_ListResources
            parameters:
                - name: id
                  in: query
                  description: The parent resource id, for example, "shelves/shelf1".
                  schema:
                    type: string
                - name: current
                  in: query
                  description: The current page number.
                  schema:
                    type: integer
                    format: int32
                - name: page_size
                  in: query
                  description: The maximum number of items to return.
                  schema:
                    type: integer
                    format: int32
                - name: page_token
                  in: query
                  description: The next_page_token value returned from a previous List request, if any.
                  schema:
                    type: string
                - name: no_paging
                  in: query
                  description: The no_paging is used to disable pagination.
                  schema:
                    type: boolean
                - name: only_count
                  in: query
                  description: The only_count is the query parameter for set only to query the total number
                  schema:
                    type: boolean
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.system.ListResourcesResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
        post:
            tags:
                - ResourceService
            operationId: ResourceService_CreateResource
            parameters:
                - name: parent
                  in: query
                  description: The parent resource id where the resource is to be created.
                  schema:
                    type: string
                - name: resource_id
                  in: query
                  description: The resource id to use for this resource.
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.v1.services.system.Resource'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.system.CreateResourceResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
    /sys/resources/{id}:
        get:
            tags:
                - ResourceService
            operationId: ResourceService_GetResource
            parameters:
                - name: id
                  in: path
                  description: |-
                    The field will contain id of the resource requested, for example:
                     "shelves/shelf1/resources/resource2"
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.system.GetResourceResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
        delete:
            tags:
                - ResourceService
            operationId: ResourceService_DeleteResource
            parameters:
                - name: id
                  in: path
                  description: |-
                    The resource id of the resource to be deleted, for example:
                     "shelves/shelf1/resources/resource2"
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.system.DeleteResourceResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
    /sys/resources/{resource.id}:
        put:
            tags:
                - ResourceService
            operationId: ResourceService_UpdateResource
            parameters:
                - name: resource.id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.v1.services.system.Resource'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.v1.services.system.UpdateResourceResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/google.rpc.Status'
...
components:
    schemas:
...
        api.v1.services.system.CreateResourceResponse:
            type: object
            properties:
                resource:
                    $ref: '#/components/schemas/api.v1.services.system.Resource'
            description: CreateResourceResponse is the response for the ResourceService.CreateResource method.
...
        api.v1.services.system.DeleteResourceResponse:
            type: object
            properties: {}
            description: DeleteResourceResponse is the response for the ResourceService.DeleteResource method.
...
        api.v1.services.system.GetResourceResponse:
            type: object
            properties:
                resource:
                    allOf:
                        - $ref: '#/components/schemas/api.v1.services.system.Resource'
                    description: The field id should match the Noun in the method id.
            description: GetResourceResponse is the response for the ResourceService.GetResource method.
...
        api.v1.services.system.ListAuthResourcesResponse:
            type: object
            properties:
                resources:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.v1.services.system.Resource'
                    description: The list of Auths.
                total_size:
                    type: integer
                    description: The total number of Auths in the result set.
                    format: int32
...
        api.v1.services.system.ListResourcesResponse:
            type: object
            properties:
                total_size:
                    type: integer
                    description: The total number of items in the list.
                    format: int32
                resources:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.v1.services.system.Resource'
                    description: The paging resources
                current:
                    type: integer
                    description: The current page number.
                    format: int32
                page_size:
                    type: integer
                    description: The maximum number of items to return.
                    format: int32
                next_page_token:
                    type: string
                    description: |-
                        Token to retrieve the next page of results, or empty if there are no
                         more results in the list.
                extra:
                    allOf:
                        - $ref: '#/components/schemas/google.protobuf.Any'
                    description: |-
                        Additional information about this response.
                         content to be added without destroying the current data format
            description: ListResourcesResponse is the response for the ResourceService.ListResources method.
...
        api.v1.services.system.Resource:
            type: object
            properties:
                id:
                    type: string
                    description: |-
                        ID of the ent.
                         field.primary_key.comment
                create_time:
                    type: string
                    description: create_time.field.comment
                    format: date-time
                update_time:
                    type: string
                    description: update_time.field.comment
                    format: date-time
                name:
                    type: string
                    description: resource.field.name
                keyword:
                    type: string
                    description: resource.field.keyword
                i_18_n_key:
                    type: string
                    description: resource.field.i18n_key
                type:
                    type: integer
                    description: resource.field.type
                    format: uint32
                status:
                    type: integer
                    description: resource.field.status
                    format: int32
                uri:
                    type: string
                    description: resource.field.uri
                operation:
                    type: string
                    description: resource.field.operation
                method:
                    type: string
                    description: resource.field.method
                component:
                    type: string
                    description: resource.field.component
                icon:
                    type: string
                    description: resource.field.icon
                sequence:
                    type: integer
                    description: resource.field.sequence
                    format: int32
                visible:
                    type: boolean
                    description: resource.field.visible
                tree_path:
                    type: string
                    description: resource.field.tree_path
                properties:
                    type: object
                    additionalProperties:
                        type: string
                    description: resource.field.properties
                description:
                    type: string
                    description: resource.field.description
                parent_id:
                    type: string
                    description: resource.field.parent_id
                children:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.v1.services.system.Resource'
                    description: Children holds the value of the children edge.
                parent:
                    allOf:
                        - $ref: '#/components/schemas/api.v1.services.system.Resource'
                    description: Parent holds the value of the parent edge.
            description: Resource is the model entity for the Resource schema.
...
        api.v1.services.system.UpdateResourceResponse:
            type: object
            properties:
                resource:
                    $ref: '#/components/schemas/api.v1.services.system.Resource'
            description: UpdateResourceResponse is the response for the ResourceService.UpdateResource method.
...
        google.protobuf.Any:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        google.rpc.Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/google.protobuf.Any'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
...
    securitySchemes:
        Basic:
            type: http
            scheme: basic
        Bearer:
            type: apiKey
            name: Authorization
            in: header
...